+++ package.json
@@ -32,6 +32,7 @@
     "node": "16"
   },
   "dependencies": {
-    "dayjs": "1.11.1"
+    "dayjs": "1.11.1",
+    "he": "1.2.0"
   }
 }
