+++ src/const.js
@@ -1,9 +1,16 @@
 const FilterType = {
   ALL: 'all',
   WATCHLIST: 'watchlist',
-  HISTORY : 'history',
+  HISTORY: 'history',
   FAVORITE: 'favorite',
 
+};
+
+const NoFilmsTextType = {
+  [FilterType.ALL]: 'There are no movies in our database',
+  [FilterType.FAVORITE]: 'There are no favorite movies now',
+  [FilterType.HISTORY]: 'There are no watched movies now',
+  [FilterType.WATCHLIST]: 'There are no movies to watch now',
 };
 
 const SortType = {
@@ -24,4 +31,4 @@ const UpdateType = {
   MAJOR: 'MAJOR',
 };
 
-export {FilterType, SortType, UserAction,UpdateType};
+export {FilterType, SortType, UserAction,UpdateType, NoFilmsTextType};
