+++ src/utils/filter.js
@@ -1,0 +1,12 @@
+import { FilterType } from '../const.js';
+
+const filter = {
+  [FilterType.ALL]: (films) => films,
+  [FilterType.WATCHLIST]: (films) => films.filter((film) => film.userDetails.watchlist),
+  [FilterType.FAVORITE]: (films) => films.filter((film) => film.userDetails.favorite),
+  [FilterType.HISTORY]: (films) => films.filter((film) => film.userDetails.alreadyWatched)
+};
+
+export {filter};
+
+
