--- src/utils/user-details.js
@@ -1,12 +1,0 @@
-import { FilterType } from '../const.js';
-
-const filter = {
-  [FilterType.ALL]:(films) => films,
-  [FilterType.WATCHLIST] : (films) => films.filter((film) => film.userDetails.watchlist),
-  [FilterType.FAVORITE] : (films) => films.filter((film) => film.userDetails.favorite),
-  [FilterType.HISTORY] : (films) => films.filter((film) => film.userDetails.alreadyWatched)
-};
-
-export {filter};
-
-
