+++ src/view/no-films-view.js
@@ -1,12 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { FilterType } from '../const.js';
-
-const NoFilmsTextType = {
-  [FilterType.ALL]: 'There are no movies in our database',
-  [FilterType.FAVORITE]: 'There are no favorite movies now',
-  [FilterType.HISTORY]: 'There are no watched movies now',
-  [FilterType.WATCHLIST]: 'There are no movies to watch now',
-};
+import {NoFilmsTextType} from '../const.js';
 
 const createNoFilmsTemplate = (filterType) => {
   const noFilmTextValue = NoFilmsTextType[filterType];
